// ssyweb/SSYNEW/src/app/claim/models/ben.model.ts
export class Beneficiary {
    benSno: number;
    benSalutation: string;
    benFirstName: string;
    benLastName: string;
    benMiddleName: string;
    benFullName: string;
    benDob: Date;
    benAge: number;
    benGender: string;
    benMaritalStatus: string;
    benFatherFirstName: string;
    benFatherLastName: string;
    benFatherMiddleName: string;
    husbandName: string;
    email: string;
    benRegDate: Date;
    benIdCardType: string;
    benIdCardNo: string;
    benReligion: string;
    benSocialCategory: string;
    benBplStatus: string;
    benBplNo: string;
    benOccupation: string;
    benMobileNo: string;
    benAddressType: string;
    benProvisionType: string;
    benProvisionNo: string;
    benEmpType: number;
    benSelfEmployed: string;
    benEstablishedAddress: string;
    benMonthlyIncom: string;
    bankDist: number;
    bankLoc: number;
    benBankName: string;
    benBankBranch: string;
    benBankAccNo: string;
    benBankIfscCode: string;
    benSsinCode: string;
    benUserId: string;
    benUserKey: string;
    benUserPwd: string;
    benIssuingAuth: string;
    benIssuingAuthName: string;
    benPhoto: string;
    benSignature: string;
    benPassBook: string;
    benIdentity: string;
    benForm: string;
    isActive: boolean;
    regNumber: string;
    regType: string;
    status: number;
    remarks: string;
    benEsiNumber: string;
    benPFStatus: number;
    aadharCard: string;
    bpicCard: string;
    bankAccNo: string;
    benAadhar: string;
    bengaliName: string;
    benTempRegNo: string;
    ifscReason: string;
    ssI_Number: number;
    iD: number;
    registrationPhotoCopy: string;
    schemeCertificate: string;
    verifyDate: Date;
    authorityLocation: string;
    salFather: string;
    salHusband: string;
    isUnique: string;
    noOfRecords: number;
    commentsByInspector: string;
    rejectedDocumnet: string;
    isDedupeDone: string;
    lastLoginDate: Date;
    loginCounts: number;
    loginStatus: number;
    reviveStatus: number;
    benOtherDoc: string;
    benDistrict: number;
    benSubDivision: number;
    benBmc: number;
    dataType: string;
    part1: string;
    part2: string;
    part2CreatedDate: Date;
    finalActionDate: Date;
    motherName: string;
    beneType: string;
    part2CreatedBy: string;
    benDeathStatus: number;
    isAssignedMultiple: boolean;
}